+++ src/__fixtures__/data/cmsSearch.ts
@@ -1,0 +1,36 @@
+export const mockCmsSearchResults = [
+  {
+    __typename: 'ArticleResult',
+    type: 'Article',
+    id: 'cl3buldda0247riyt6h85cpgc',
+    permalink: 'my-fitness-article',
+    title: 'My Fitness Article',
+    preview: 'This is the preview text!',
+    date: '2022-05-18T17:18:40.802Z',
+    labels: [
+      {
+        __typename: 'Label',
+        id: 'cl4fqrdce0451u5a8s592596k',
+        name: 'All Guardians',
+      },
+    ],
+  },
+  {
+    __typename: 'BookmarkResult',
+    type: 'Bookmark',
+    id: 'cl3bpuzvq0885i6ytzr2oumu9',
+    permalink: 'https://myfss.us.af.mil/',
+    title: 'MyFitness',
+    preview:
+      'Your destination for getting notifications, scheduling, entering test results, and all other processes related to the fitness goals for Airmen and Guardians.',
+  },
+  {
+    __typename: 'BookmarkResult',
+    id: 'cktd7e2tr0358w597t5j24os9',
+    type: 'Bookmark',
+    title: 'PT Tests',
+    preview: 'View your PT Test scores.',
+    permalink:
+      'https://myfss.us.af.mil/USAFCommunity/s/usaf-fitness-management',
+  },
+]