+++ src/lib/mongodb.ts
@@ -7,11 +7,11 @@ import { MongoClient } from 'mongodb'
 // Required to use global._mongoClientPromise
 declare global {
   /* eslint-disable-next-line */
-  namespace NodeJS {
-    interface Global {
-      _mongoClientPromise: Promise<typeof MongoClient>
-    }
-  }
+  var _mongoClientPromise: Promise<typeof MongoClient>
+}
+
+const globalWithMongo = global as typeof globalThis & {
+  _mongoClientPromise: Promise<typeof MongoClient>
 }
 
 const host = process.env.MONGO_HOST || ''
@@ -31,11 +31,11 @@ let clientPromise: Promise<typeof MongoClient>
 if (process.env.NODE_ENV === 'development') {
   // In development mode, use a global variable so that the value
   // is preserved across module reloads caused by HMR (Hot Module Replacement).
-  if (!global._mongoClientPromise) {
+  if (!globalWithMongo._mongoClientPromise) {
     client = new MongoClient(connectionString)
-    global._mongoClientPromise = client.connect()
+    globalWithMongo._mongoClientPromise = client.connect()
   }
-  clientPromise = global._mongoClientPromise
+  clientPromise = globalWithMongo._mongoClientPromise
 } else {
   // In production mode, it's best to not use a global variable.
   client = new MongoClient(connectionString)